+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @expect/browser
 
+## 0.0.25
+
+### Patch Changes
+
+- fix
+- fix
+- Updated dependencies
+- Updated dependencies
+  - @expect/cookies@0.0.25
+  - @expect/shared@0.0.25
+
 ## 0.0.24
 
 ### Patch Changes